Fix issue with broken search, small refactor
